Commit 2
Changed to committing to the Updates branch. Makes a lot more sense.

New changes from last time:
- Added Treants to the Forest map, and removed the 'Boss' enemy. Treants
can move through trees.
- Added an Armour icon to the HUD, next to the Sword.
- Added a whole new selection of loot called 'ExtraLoot'. The primary
function of this was so I could make Lumber Axes always spawn once (and
only once) on Forest maps, but this could easily be used for other
unique loot on different maps. Currently, it just spawns a cloned
version of the Wand and the Health Potion.
- There is commented out code that shows my attempt at making it so you
couldn't pick up Health or Speed potions if you were at max Health/Speed
(respectively). The game kept crashing. Left it in for when I can make
it work.
- The Health graphic has been reduced to 8x8 pixels. The idea is that
this will let it stack better. However, that won't /really/ show until I
make progression work.
- Added Speed potions, but in this commit they don't do anything - the
code is all commented out. They SEEM to work, but I can't get the HUD
element to increase as the player picks up a potion. Traps (which take
away Speed, both in-game and on the HUD) seem to work just fine, not
sure why this doesn't.